almost done

+++ movieRecomander/backend/src/main/java/com/movierecommender/backend/comments/CommentController.java
@@ -1,6 +1,8 @@
 package com.movierecommender.backend.comments;
 
 import com.movierecommender.backend.advice.BusinessException;
+import com.movierecommender.backend.identity.IdentityService;
+import com.movierecommender.backend.security.facades.AuthenticationFacade;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,10 +15,12 @@ import java.util.List;
 @RequestMapping("api/v1/comments")
 public class CommentController {
     private final CommentRepository commentRepository;
+    private final IdentityService identityService;
 
     @Autowired
-    public CommentController(CommentRepository commentRepository) {
+    public CommentController(CommentRepository commentRepository, IdentityService identityService) {
         this.commentRepository = commentRepository;
+        this.identityService = identityService;
     }
 
     @GetMapping
@@ -43,24 +47,36 @@ public class CommentController {
     }
 
     @PutMapping("/{id}")
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN')")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_USER')")
     @ResponseStatus(code = HttpStatus.NO_CONTENT, reason = "UPDATED")
     public void update(@PathVariable("id") Long id, @RequestBody Comment comment) {
         var foundComment = commentRepository.findById(id);
         if (foundComment.isEmpty()) {
             throw new BusinessException("Comment not found", "Invalid data", HttpStatus.NOT_FOUND);
         }
+
+        var currentUser = this.identityService.getLoggedInUser();
+        if (currentUser.isEmpty() || !currentUser.get().equals(foundComment.get().getAppUser())) {
+            throw new BusinessException("User can only update his own comment", "Invalid permission", HttpStatus.FORBIDDEN);
+        }
+
         foundComment.get().update(comment);
     }
 
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN')")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_USER')")
     @ResponseStatus(code = HttpStatus.NO_CONTENT, reason = "DELETED")
     public void delete(@PathVariable("id") Long id) {
         var foundComment = commentRepository.findById(id);
         if (foundComment.isEmpty()) {
             throw new BusinessException("Comment not found", "Invalid data", HttpStatus.NOT_FOUND);
         }
+
+        var currentUser = this.identityService.getLoggedInUser();
+        if (currentUser.isEmpty() || !currentUser.get().equals(foundComment.get().getAppUser())) {
+            throw new BusinessException("User can only delete his own comment", "Invalid permission", HttpStatus.FORBIDDEN);
+        }
+
         commentRepository.delete(foundComment.get());
     }
 }
